Updated the HandleSliceUpdate for the various geometry layers. The publisher sends the Bytebuffer with the Objectid last. This originally caused a memory leak for the geometry object.
This mod places the datahandling after the parsing while loop of the bytebuffer. When multiple objects are in the bytebuffer, the new code only add the last object. The old code did not handle correctly either.

When a geometry is present in the bytebuffer, and the id was already known, the new code replaces the geometry object with the new one.