weatherControlView.js - renamed version for WindTemperatureView.js along with minor fixes weatherControlView.css - renamed version for WindTemperatureView.css
WindDirection.js - Added functionality to both act alone and along with the temperature control.

temperatureControlView.js - Improved logic for units (used the units defined in the UnitConverter instead of defining a separate object)
temperatureControlModel.js - Improved logic for units (used the units defined in the UnitConverter instead of defining a separate object). Also reduced calculations when a new unit is set.

General updates:
* Removed unused code
* Used Lower Camel Case style for naming files
* Removed redundant code